Make the game clock more accurate
There's no guarantee that a timeout will finish on time, and after
a short period the clock becomes inaccurate.
With the periodic refresh this isn't a big deal, but the countdown
doesn't look as smooth as it could.
By keeping track of the actual end date there will be no jumps,
and it looks smoother overall.

This also makes the code much easier to understand & maintain.